refactor(all)!: adopt vanilla Kiota request-configuration pattern (#511)
* refactor(all)!: adopt vanilla Kiota request-configuration pattern

Converts every module's *QueryParameters structs to pointer fields with
uriparametername tags and wires request builders through
abstractions.ConfigureRequestInformation directly, matching the
Kiota-generated SDK convention (e.g. msgraph-sdk-go) instead of the
repo's own go-querystring-based wrapper. Resolves the spike from #494:
the "standard pattern" question is breaking, not additive.

Removes the now-dead internal.ConfigureRequestInformation /
internal.KiotaRequestInformation wrapper and drops the go-querystring
dependency entirely.

BREAKING CHANGE: every exported *QueryParameters struct across every
API module now uses pointer fields (e.g. `Limit int` -> `Limit *int32`,
`Query string` -> `Query *string`, `DisplayValue DisplayValue` ->
`DisplayValue *DisplayValue`) instead of value types. Callers building
these structs with value literals must switch to pointers (see
internal.ToPointer).


* chore: add design-decisions skill and ignore docusaurus artifacts

Adds a design-decisions Claude skill pointing at this repo's docs/adr/
ADRs (error standardization, backing-store models, hand-written-on-Kiota),
and extends .gitignore to exclude website/.docusaurus and
website/node_modules build artifacts.


* fix(all): update build-tag-gated code for pointer query parameters

Fixes website/snippets (go vet -tags snippets, run by .github/workflows/docs.yml)
and tests/integration (go test -tags integration, run by ci.yml), both of which
were missed by the previous commit's test plan since neither is covered by a
plain `go build`/`go test ./...`. Also adds an end-to-end regression test for
tableapi's GET query encoding through a baseurl-based builder (the shape the
real client uses), covering the pointer/uriparametername-tag conversion with
an assertion on the actual encoded URL rather than just the HTTP method.


---------

+++ website/snippets/attachments_recent.go
@@ -32,7 +32,7 @@ func _() {
 	}
 
 	// Step 3: List attachments (filter using query params as needed)
-	limit := 1
+	limit := int32(1)
 	encodedQuery := "ORDERBYDESCsys_created_on"
 
 	config := &attachmentapi.AttachmentRequestBuilderGetRequestConfiguration{
